% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_wrapper.R
\name{.close_remDr}
\alias{.close_remDr}
\title{Close to connect remote server}
\usage{
.close_remDr(remDr)
}
\arguments{
\item{remDr}{Object remote driver}
}
\description{
Close to connect remote server
}
