% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_wrapper.R
\name{.close_remote_driver}
\alias{.close_remote_driver}
\title{Close to connect remote server}
\usage{
.close_remote_driver(remote_driver)
}
\arguments{
\item{remote_driver}{Object remote driver}
}
\description{
Close to connect remote server
}
