% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_scraper.R
\name{.create_album_reviews_df}
\alias{.create_album_reviews_df}
\title{Create dataframe with all the album's critic reviews}
\usage{
.create_album_reviews_df(obj_scrap)
}
\arguments{
\item{obj_scrap}{Object scrapper}
}
\value{
Dataframe with reviews
}
\description{
Create dataframe with all the album's critic reviews
}
