% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_scraper.R
\name{.create_best_albums_df}
\alias{.create_best_albums_df}
\title{Create dataframe with top 100 best albums}
\usage{
.create_best_albums_df(obj_scrap)
}
\arguments{
\item{obj_scrap}{Object scrapper}
}
\value{
Dataframe with best albums
}
\description{
Create dataframe with top 100 best albums
}
