% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_name}
\alias{.format_name}
\title{Format name to url's pattern}
\usage{
.format_name(name)
}
\arguments{
\item{name}{Name to format}
}
\value{
Formated name
}
\description{
Format name to url's pattern
}
