% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_wrapper.R
\name{.open_remote_driver}
\alias{.open_remote_driver}
\title{Open to connect to remote server}
\usage{
.open_remote_driver()
}
\value{
remote_driver Object remote driver
}
\description{
Open to connect to remote server
}
