% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_scraper.R
\name{.scrape_album_critic_reviews}
\alias{.scrape_album_critic_reviews}
\title{Extract album reviews}
\usage{
.scrape_album_critic_reviews(remDr, name, artist)
}
\arguments{
\item{remDr}{Remote driver}

\item{name}{Album's name}

\item{name}{Album's autor}
}
\value{
Return list with object of scrapper
}
\description{
Extract album reviews
}
