% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_scraper.R
\name{get_album_critic_reviews}
\alias{get_album_critic_reviews}
\title{Get album critic reviews}
\usage{
get_album_critic_reviews(name, artist)
}
\arguments{
\item{name}{Album's name}

\item{artist}{Album's autor}
}
\value{
Dataframe containing all album's critic reviews
}
\description{
Get album critic reviews
}
\examples{
album_critics <- get_album_critic_reviews("Melodrama", "Lorde")
}
