% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_scraper.R
\name{get_best_albums_per_year}
\alias{get_best_albums_per_year}
\title{Get best metacritic albums}
\usage{
get_best_albums_per_year(year, by = "metascore")
}
\arguments{
\item{year}{Year of albums}

\item{by}{Selector}
}
\value{
Dataframe containing the best 100 albums of the year
}
\description{
Get best metacritic albums
}
\examples{
best_2018_albums <- get_best_albums_per_year("2018")
best_2018_albums <- get_best_albums_per_year("2018", by = "shared")
}
