% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_scraper.R
\name{.scrape_best_albums_per_year}
\alias{.scrape_best_albums_per_year}
\title{Extract best albums by year}
\usage{
.scrape_best_albums_per_year(remDr, year, url)
}
\arguments{
\item{remDr}{Remote driver}

\item{year}{Year of albums}
}
\value{
Return list with object of scrapper
}
\description{
Extract best albums by year
}
